Add direct link to download the slides
If you right click and try to save the PDF file through GitHub it ends
up being a corrupt PDF. You would have to click the PDF and then use the
explicit download button.

This link bypasses having to do that and allows you to directly download
it from GitHub.